Refactored CSV
Fixed the warnings CSV_Util and Table
Fixed Copy Paste Error xD

+++ src/utils/CSV_Util.h
@@ -31,7 +31,7 @@ public:
 
 	static float strToFloat(const std::string element);
 
-	static float strToDouble(const std::string element);
+	static double strToDouble(const std::string element);
 	
 private:
 	// Contains the error string describing failed operation
